requirements.txt
streamlit
pandas
numpy
sentence-transformers
faiss-cpu
wikipedia-api
beautifulsoup4
langchain-groq
langchain-core
langchain